# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for SMIL",
        "contact": {
            "email": "muhammad.raflysadewa.tik17@mhsw.pnj.ac.id"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/image-alat/get-by-alat-id/{alatId}": {
            "get": {
                "tags": [
                    "Image Alat"
                ],
                "summary": "Return List of Image Alat By Id",
                "operationId": "getAllImageAlat",
                "parameters": [
                    {
                        "name": "alatId",
                        "in": "path",
                        "description": "Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/image-alat": {
            "post": {
                "tags": [
                    "Image Alat"
                ],
                "summary": "Create New ImageAlat",
                "operationId": "createNewImageAlat",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreImageAlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageAlatListResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/image-alat/{id}": {
            "delete": {
                "tags": [
                    "Image Alat"
                ],
                "summary": "Delete Existed Image Alat",
                "operationId": "deleteExistingImageAlat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ImageAlat": {
                "title": "ImageAlat",
                "description": "ImageAlat Model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "alat_id": {
                        "title": "Alat ID",
                        "description": "ID of Alat",
                        "type": "integer"
                    },
                    "alat_image_model": {
                        "title": "Alat Model",
                        "description": "Model of Alat"
                    },
                    "image_data": {
                        "title": "Image Data",
                        "description": "Data of Image in base64 Format",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created Date",
                        "description": "Date of data been created",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ImageAlat"
                }
            },
...
            "Page": {
                "title": "Page",
                "description": "Page Model",
                "properties": {
                    "current": {
                        "title": "Current",
                        "description": "Current Page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "total": {
                        "title": "Total Page",
                        "description": "Total Page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "size": {
                        "title": "Size Page",
                        "description": "Total Data display in Page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "data_total": {
                        "title": "Data Total",
                        "description": "Total Data store in system",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Page"
                }
            },
...
            "StoreImageAlatRequest": {
                "title": "Store Image Alat Request",
                "description": "Store Image Alat Request body data",
                "required": [
                    "image_data_list",
                    "alat_id"
                ],
                "properties": {
                    "alat_id": {
                        "title": "alat_id",
                        "description": "ImageAlat Name",
                        "type": "integer"
                    },
                    "image_data_list": {
                        "title": "image_data_list",
                        "description": "Image Alat List",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "ImageAlatListResource": {
                "title": "ImageAlatListResource",
                "description": "ImageAlat Resource",
                "properties": {
                    "result": {
                        "title": "Result",
                        "description": "Result ImageAlat Wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImageAlat"
                        }
                    },
                    "page": {
                        "$ref": "#/components/schemas/Page"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AlatListResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}